requirements.txt
line-bot-sdk
flask
requests
pandas
numpy
beautifulsoup4
gspread
oauth2client
lxml
